+++ src/parser/block/paragraph/mention/address.test.ts
@@ -22,22 +22,22 @@ describe('Unit: parser/block/paragraph/mention/address', () => {
     });
 
     it('valid', () => {
-      assert.deepStrictEqual(inspect(parser('>0')), [['<a rel="noopener" class="address" data-level="1">&gt;0</a>'], '']);
-      assert.deepStrictEqual(inspect(parser('>a')), [['<a rel="noopener" class="address" data-level="1">&gt;a</a>'], '']);
-      assert.deepStrictEqual(inspect(parser('>a ')), [['<a rel="noopener" class="address" data-level="1">&gt;a</a>'], '']);
-      assert.deepStrictEqual(inspect(parser('>a\n')), [['<a rel="noopener" class="address" data-level="1">&gt;a</a>'], '']);
-      assert.deepStrictEqual(inspect(parser('>A')), [[`<a rel="noopener" class="address" data-level="1">&gt;A</a>`], '']);
-      assert.deepStrictEqual(inspect(parser('>>0')), [['<a rel="noopener" class="address" data-level="2">&gt;&gt;0</a>'], '']);
-      assert.deepStrictEqual(inspect(parser('>0\n1')), [['<a rel="noopener" class="address" data-level="1">&gt;0</a>'], '1']);
-      assert.deepStrictEqual(inspect(parser('>0\n>')), [['<a rel="noopener" class="address" data-level="1">&gt;0</a>'], '>']);
-      assert.deepStrictEqual(inspect(parser('>0\n>1')), [['<a rel="noopener" class="address" data-level="1">&gt;0</a>', '<a rel="noopener" class="address" data-level="1">&gt;1</a>'], '']);
-      assert.deepStrictEqual(inspect(parser('>0\n>>1')), [['<a rel="noopener" class="address" data-level="1">&gt;0</a>', '<a rel="noopener" class="address" data-level="2">&gt;&gt;1</a>'], '']);
-      assert.deepStrictEqual(inspect(parser('>>0\n>1')), [['<a rel="noopener" class="address" data-level="2">&gt;&gt;0</a>', '<a rel="noopener" class="address" data-level="1">&gt;1</a>'], '']);
-      assert.deepStrictEqual(inspect(parser('>https://host?<')), [['<a rel="noopener" target="_blank" class="address" data-level="1">&gt;https://host?&lt;</a>'], '']);
+      assert.deepStrictEqual(inspect(parser('>0')), [['<a href="?log=0" rel="noopener" class="address" data-level="1">&gt;0</a>'], '']);
+      assert.deepStrictEqual(inspect(parser('>a')), [['<a href="?log=a" rel="noopener" class="address" data-level="1">&gt;a</a>'], '']);
+      assert.deepStrictEqual(inspect(parser('>a ')), [['<a href="?log=a" rel="noopener" class="address" data-level="1">&gt;a</a>'], '']);
+      assert.deepStrictEqual(inspect(parser('>a\n')), [['<a href="?log=a" rel="noopener" class="address" data-level="1">&gt;a</a>'], '']);
+      assert.deepStrictEqual(inspect(parser('>A')), [[`<a href="?log=A" rel="noopener" class="address" data-level="1">&gt;A</a>`], '']);
+      assert.deepStrictEqual(inspect(parser('>>0')), [['<a href="?log=0" rel="noopener" class="address" data-level="2">&gt;&gt;0</a>'], '']);
+      assert.deepStrictEqual(inspect(parser('>0\n1')), [['<a href="?log=0" rel="noopener" class="address" data-level="1">&gt;0</a>'], '1']);
+      assert.deepStrictEqual(inspect(parser('>0\n>')), [['<a href="?log=0" rel="noopener" class="address" data-level="1">&gt;0</a>'], '>']);
+      assert.deepStrictEqual(inspect(parser('>0\n>1')), [['<a href="?log=0" rel="noopener" class="address" data-level="1">&gt;0</a>', '<a href="?log=1" rel="noopener" class="address" data-level="1">&gt;1</a>'], '']);
+      assert.deepStrictEqual(inspect(parser('>0\n>>1')), [['<a href="?log=0" rel="noopener" class="address" data-level="1">&gt;0</a>', '<a href="?log=1" rel="noopener" class="address" data-level="2">&gt;&gt;1</a>'], '']);
+      assert.deepStrictEqual(inspect(parser('>>0\n>1')), [['<a href="?log=0" rel="noopener" class="address" data-level="2">&gt;&gt;0</a>', '<a href="?log=1" rel="noopener" class="address" data-level="1">&gt;1</a>'], '']);
+      assert.deepStrictEqual(inspect(parser('>https://host?<')), [['<a href="https://host?<" rel="noopener" target="_blank" class="address" data-level="1">&gt;https://host?&lt;</a>'], '']);
     });
 
     it('nofollow', () => {
-      assert.deepStrictEqual(inspect(parser('>ttps://host')), [['<a rel="noopener nofollow noreferrer" target="_blank" class="address" data-level="1">&gt;ttps://host</a>'], '']);
+      assert.deepStrictEqual(inspect(parser('>ttps://host')), [['<a href="https://host" rel="noopener nofollow noreferrer" target="_blank" class="address" data-level="1">&gt;ttps://host</a>'], '']);
     });
 
   });